RATIS-1168. Move the utils from DataStreamBaseTest to a new file.
